scripts/generate_file/requirements.txt
faker
python-docx
openpyxl
reportlab
python-pptx
tqdm
pandas